Move TaskHelper to Util.h
