+++ source/yats/include/yats/Util.h
@@ -16,6 +16,7 @@ struct has_run
 template<typename T>
 constexpr bool has_run_v = has_run<T>::value;
 
+
 template<typename T>
 struct is_unique_ptr
 {
@@ -33,6 +34,7 @@ struct is_unique_ptr<void>
 template<typename T>
 constexpr bool is_unique_ptr_v = is_unique_ptr<T>::value;
 
+
 template<typename T>
 struct is_shared_ptr
 {
@@ -49,5 +51,22 @@ struct is_shared_ptr<void>
 
 template<typename T>
 constexpr bool is_shared_ptr_v = is_shared_ptr<T>::value;
+
+
+template <typename Return, typename... ParameterTypes>
+struct TaskHelper
+{
+	template <typename CompoundType>
+	static typename CompoundType::value_type transform();
+
+	using WrappedInput = std::tuple<ParameterTypes...>;
+	using Input = std::tuple<decltype(transform<ParameterTypes>())...>;
+	using ReturnType = Return;
+
+	static constexpr size_t ParameterCount = sizeof...(ParameterTypes);
+};
+
+template <typename ReturnType, typename TaskType, typename... ParameterTypes>
+static constexpr TaskHelper<ReturnType, ParameterTypes...> MakeHelper(ReturnType(TaskType::*)(ParameterTypes...));
 
 }  // namespace yats
